Separated IStats from StatsController
StatsController created to organize thinsg better
