+++ TouchCrawlerUnity/Assets/Scripts/Stats/StatsController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 217aa8c8037a4f74b81412d4c5cfb3ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
